custom_agent/requirements.txt
fastapi
uvicorn
crewai
pinecone
google-generativeai
requests
langchain-openai
python-dotenv
asyncio